Tests helps convert work better
The name did not match the required specification.

Convert as a method name is terrible, so we provide an alias to that,
while using the better named f_to_c method name.

Checked for edge cases.

Found and fixed a bug due to testing.  Will no longer do "integer
division".